Update adding-a-new-ssh-key-to-your-github-account.md
The GitHub CLI does not allow you to add an ssh key unless you are authenticated with your GitHub account. The `gh auth login` command not only authenticates you but also gives prompts to add your ssh key to your account. This PR amends the GitHub CLI portion with the `gh auth login` path. This may be useful to someone coming to this page for the first time who has not already installed the gh CLI and will do so as part of this flow and will then need to authenticate.